+++ glean-core/android/src/test/java/mozilla/telemetry/glean/GleanTest.kt
@@ -484,6 +484,7 @@ class GleanTest {
             schedulesPings = emptyList(),
             reasonCodes = listOf(),
             followsCollectionEnabled = true,
+            uploaderCapabilities = emptyList(),
         )
         val stringMetric = StringMetricType(
             CommonMetricData(
@@ -555,11 +556,11 @@ class GleanTest {
     }
 
     @Test
-    fun `Core metrics should be cleared and restored when disabling and enabling uploading`() {
+    fun `Core metrics are not cleared when disabling and enabling uploading`() {
         assertNotNull(GleanInternalMetrics.os.testGetValue())
 
         Glean.setCollectionEnabled(false)
-        assertNull(GleanInternalMetrics.os.testGetValue())
+        assertNotNull(GleanInternalMetrics.os.testGetValue())
 
         Glean.setCollectionEnabled(true)
         assertNotNull(GleanInternalMetrics.os.testGetValue())
@@ -855,6 +856,7 @@ class GleanTest {
             schedulesPings = emptyList(),
             reasonCodes = listOf(),
             followsCollectionEnabled = true,
+            uploaderCapabilities = emptyList(),
         )
         val stringMetric = StringMetricType(
             CommonMetricData(