+++ glean-core/rlb/examples/prototype.rs
@@ -39,6 +39,7 @@ pub static PrototypePing: Lazy<PingType> = Lazy::new(|| {
         vec![],
         vec![],
         true,
+        vec![],
     )
 });